web-animations/animation-player/animation-player-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer currentTime attribute tests</title>
<meta name="assert" content="The current time of this player unless this player has a pending pause task, in which case this attribute returns null">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationplayer-currenttime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
var timesToTest = [-100,
    -1,
    0,
    1,
    100,
    ANIMATION_END_TIME / 2,
    ANIMATION_END_TIME - 100,
    ANIMATION_END_TIME,
    ANIMATION_END_TIME + 1,
    ANIMATION_END_TIME + 100];

function calculateCurrentTime(timelineTime, startTime, playbackRate) {
    return (timelineTime - startTime) * playbackRate;
}

function calculateStartTime(timelineTime, seekTime, playbackRate) {
    return timelineTime - seekTime / playbackRate;
}

test(function() {
    var player = new AnimationPlayer();
    assert_equals(player.currentTime, null, 'AnimationPlayer.currentTime should be null');
}, 'Test that AnimationPlayer.currentTime attribute is null if player has no associated timeline');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));

    player.ready.then(t.step_func(function() {
        assert_equals(player.currentTime, calculateCurrentTime(
            player.timeline.currentTime, player.startTime, player.playbackRate),
            'AnimationPlayer.currentTime value should be calculated as defined');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = 0.4;

    player.ready.then(t.step_func(function() {
        setTimeout(function() {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be calculated as defined');
            t.done();
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate. ' +
    'Test 0 < playbackRate < 1');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = 2;

    player.ready.then(t.step_func(function() {
        setTimeout(function() {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be calculated as defined');
            t.done();
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate. ' +
    'Test playbackRate > 1');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.currentTime = ANIMATION_END_TIME / 2;
    player.playbackRate = -1.5;

    player.ready.then(t.step_func(function() {
        setTimeout(function() {
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be calculated as defined');
            t.done();
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.currentTime value is (timeline time - start time) × playback rate. ' +
    'Test playbackRate < 0');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    assert_throws('NotSupportedError', function() {
        player.currentTime = null;
    }, 'NotSupportedError should be thrown if value to set is unresolved');
}, 'Setting AnimationPlayer.currentTime to unresolved value throws NotSupportedError');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));

    player.ready.then(t.step_func(function() {
        timesToTest.forEach(function(val) {
            player.currentTime = val;
            assert_equals(player.currentTime, val,
                'AnimationPlayer.currentTime value should be set to the new value');
            assert_equals(player.currentTime, calculateCurrentTime(
                player.timeline.currentTime, player.startTime, player.playbackRate),
                'AnimationPlayer.currentTime value should be calculated as defined');
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test setting value of AnimationPlayer.currentTime attribute');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));

    player.ready.then(t.step_func(function() {
        timesToTest.forEach(function(seekTime) {
            var expectedTime = calculateStartTime(player.timeline.currentTime, seekTime,
                player.playbackRate);
            player.currentTime = seekTime;

            assert_equals(player.startTime, expectedTime,
                'The value of AnimationPlayer.startTime attribute should be set when seek ' +
                'is performed');
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.currentTime attribute sets start time');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    var playbackRates = [-2, -0.4, 0.4, 2];

    player.ready.then(t.step_func(function() {
        playbackRates.forEach(function(playbackRate) {
            player.playbackRate = playbackRate;
            timesToTest.forEach(function(seekTime) {
                var expectedTime = calculateStartTime(player.timeline.currentTime, seekTime,
                    player.playbackRate);
                player.currentTime = seekTime;

                assert_equals(player.startTime, expectedTime,
                    'The value of AnimationPlayer.startTime attribute should be set when ' +
                    'seek is performed for playbackRate ' + playbackRate);
            });
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.currentTime attribute correctly ' +
    'sets start time for different playback rates');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = 0;

    player.ready.then(t.step_func(function() {
        var startTime = player.startTime;
        timesToTest.forEach(function(seekTime) {
            player.currentTime = seekTime;
            assert_equals(player.startTime, startTime,
                'The value of AnimationPlayer.startTime attribute should not be updated');
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.currentTime attribute does not ' +
    'change start time if playback rate is 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.pause();

    player.ready.then(t.step_func(function() {
        var startTime = player.startTime;
        timesToTest.forEach(function(seekTime) {
            player.currentTime = seekTime;
            assert_equals(player.startTime, startTime,
                'The value of AnimationPlayer.startTime attribute should not be updated');
            t.done();
        });
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.currentTime attribute does not ' +
    'change start time if player hold time is resolved');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));

    player.ready.then(t.step_func(function() {
        player.currentTime = ANIMATION_END_TIME + 1;
        assert_equals(player.playState, 'finished', 'Player state should be finished');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.currentTime attribute to time after end ' +
    'finishes the player');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = -1;

    player.ready.then(t.step_func(function() {
        player.currentTime = -1;
        assert_equals(player.playState, 'finished', 'Player state should be finished');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that setting value of AnimationPlayer.currentTime attribute to negative ' +
    'value finishes player if playback rate < 0');

async_test(function(t) {
    var animation = newAnimation(createDiv(t));
    var player = document.timeline.play(animation);

    player.ready.then(t.step_func(function() {
        player.currentTime = 0;
        assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0,
            'Animation effect should take place');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time is zero');

async_test(function(t) {
    var animation = newAnimation(createDiv(t));
    var player = document.timeline.play(animation);

    player.ready.then(t.step_func(function() {
        player.currentTime = ANIMATION_END_TIME / 2;
        assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0_5,
            'Animation effect should take place');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time is > 0 and < end time');

async_test(function(t) {
    var animation = newAnimation(createDiv(t));
    var player = document.timeline.play(animation);

    player.ready.then(t.step_func(function() {
        player.currentTime = ANIMATION_END_TIME;
        assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_1,
            'Animation effect should take place');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time equals end time');

// TODO Add tests for setting current time when player has no timeline associated
</script>
</body>
